+++ CHANGES.md
@@ -3,6 +3,13 @@ Pykit Changelog
 
 Here you can see the full list of changes between each Pykit release.
 
+Version 0.5.0
+-----------
+- fixed: initializing table with schema with data resulted in error
+- fixed: table `align` now is also applied to header
+- fixed: `closeButton` not working on modal
+- changed: `__after__` and `__init__` call order reversed, ancestor -> child
+
 Version 0.3.12
 -----------
 - changed: dark theme color tweaks